BRESPI - BRESENHAM AND PI CIRCLES IN COLOURS OF R G B
BRESPI - BRESENHAM AND PI CIRCLES IN COLOURS OF R G B

This app is drawing circles in graphics mode. The aim is to compare which value of Pi give the best results with covering circles drawn with Bresenham method. The App is treating the screen in chosen resolution as matrix and counting pixels. Dependencies: Usage of Allegro 4.2 library, by Shawn Hargreaves.